+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Notificaiton subscriptions - Shows GiB unit for network in modal
+- Enables bulk actions in table plugin - Moves deployments bulk action feature to table plugin, #1180
+
 ## [2.33.3] - 2023-02-27
 
 ### Changed